pyproject.toml
[tool.poetry]
name = "service-length"
version = "0.1.0"
description = ""
authors = ["zre12f"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.0.7"
psycopg = "3.2.1"
psycopg-binary = "3.2.1"
python-dateutil = "2.9.0.post0"
six = "1.16.0"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
